// web/src/lib/api/generated/models/githubComNgodingvarengMemoriaInternalDeliveryRestDtoThreadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Book API
 * OpenAPI spec version: 1.0
 */

export interface GithubComNgodingvarengMemoriaInternalDeliveryRestDtoThreadResponse {
  archived_at?: string;
  circle_id?: string;
  color_hex?: string;
  created_at?: string;
  description?: string;
  id?: string;
  name?: string;
  sort_order?: number;
  updated_at?: string;
  user_id?: string;
}